requirements.txt
pyTelegramBotAPI==4.15.2
python-dotenv==1.0.0
CurrencyConverter==0.17.15
